requirements.txt
tqdm
numpy
matplotlib
fsspec[http]>=2021.05.0, !=2021.06.0
joblib
hydra_core==1.1.1
hydra_colorlog==1.1
hydra-joblib-launcher==1.1.5
torchdiffeq
seaborn
plotly
libigl
pyvista